requirements.txt
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
typer>=0.9.0
rich>=13.0.0
email-validator>=2.1.0
python-dateutil>=2.8.2
psutil>=5.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
coverage>=7.3.0

# HTML parsing and email cleaning
beautifulsoup4>=4.12.0
html2text>=2020.1.16
email-reply-parser>=0.5.12

# Logging
loguru>=0.7.0

# CLI testing
click>=8.1.0
